refacto(validator): integrate utils pkg as it is only used by validator
