+++ pkg/validator/utils_test.go
@@ -1,4 +1,4 @@
-package utils
+package validator
 
 import (
 	"os"
@@ -28,7 +28,7 @@ func TestCheckUsernameFormat(t *testing.T) {
 	}
 	for _, test := range tests {
 		t.Run(test.testName, func(t *testing.T) {
-			res := CheckUsernameFormat(test.username)
+			res := checkUsernameFormat(test.username)
 			assert.Equal(t, test.res, res)
 		})
 	}
@@ -36,19 +36,19 @@ func TestCheckUsernameFormat(t *testing.T) {
 
 func TestCheckNotWritable(t *testing.T) {
 	t.Run("File does not exist - false", func(t *testing.T) {
-		res := CheckNotWritable("MISSING")
+		res := checkNotWritable("MISSING")
 		assert.False(t, res)
 	})
 }
 
 func TestGetEnv(t *testing.T) {
 	t.Run("Env var does not exist - falback", func(t *testing.T) {
-		res := GetEnv("THIS_ENV_VER_DOES_NOT_EXIST", "value")
+		res := getEnv("THIS_ENV_VER_DOES_NOT_EXIST", "value")
 		assert.Equal(t, res, "value")
 	})
 	t.Run("Env var is empty - falback", func(t *testing.T) {
 		_ = os.Setenv("THIS_ENV_VAR_IS_EMPTY", "")
-		res := GetEnv("THIS_ENV_VAR_IS_EMPTY", "value")
+		res := getEnv("THIS_ENV_VAR_IS_EMPTY", "value")
 		_ = os.Unsetenv("THIS_ENV_VAR_IS_EMPTY")
 		assert.Equal(t, res, "value")
 	})
